+++ frontend/src/App.jsx
@@ -1,13 +1,11 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
 import './App.css'
+import LandingNavbar from './components/landing/LandingNavbar.jsx'
 
 function App() {
   return (
     <div className='container text-center mt-5'>
-      <h1 className="text-primary">Welcome to Vite + React</h1>
-      <button className='btn btn-success'>Click Me</button>
+      <LandingNavbar />
     </div>
   )
 }